Update README.md
Latest-Stable-Version-23.8.16-M2
23.8.16 - Date
M2 - Marker Based Watershed Segmentation.

This is the latest stable version. By stable version it means that this version was both compiling and executing successfully. In this version the ROI was being selected successfully by the user. The seed point is also being selected but it has no been used to generate the marker image. The marker image is being generated automatically.